Splitting Quartile & Entropy discretizers
